* Floating point exceptions of some sensible types will now be raised and   maybe cause some program terminations where there were no problems before.   This is so that we can track bugs earlier, and we can still disable this   setting shortly before the next release to ensure smooth gameplay. * Fixed the bug in the menu code that would cause a justified and horrible   floating point exception.  Justified because uninitialized memory was   used in a computation.

+++ src/misc.c
@@ -1601,7 +1601,7 @@ CheckForTriggeredEventsAndStatements ( int PlayerNum )
   //--------------------
   // Now we check if some statment location is reached
   //
-  map_x = (int) rintf( Me [ PlayerNum ] . pos . x ); map_y = (int) rintf( Me [ PlayerNum ] . pos . y ) ;
+  map_x = (int) rintf( (float) Me [ PlayerNum ] . pos . x ); map_y = (int) rintf( (float)Me [ PlayerNum ] . pos . y ) ;
   for ( i = 0 ; i < MAX_STATEMENTS_PER_LEVEL ; i++ )
     {
       if ( ( map_x == StatementLevel -> StatementList [ i ] . x ) &&
